refactor: audit _all_ unsafe blocks
Now I feel comfortable to continue doing changes, since the API should now be
sound. It would be nice to remove all the unsafe code all together, but I think
this is the best that I could do without creating more safer bindings for
`rusty_spine`.

There is some wording that could be improved, but for now, this is good enough
. :)